backend/requirements.txt
# Core ML dependencies
torch>=2.2.0
torchvision>=0.17.0
accelerate>=0.26.0
transformers>=4.37.0

# Image generation dependencies
diffusers>=0.27.0
Pillow>=10.0.0
numpy>=1.24.0

# Model format handlers
safetensors>=0.4.0
huggingface-hub>=0.19.0
ctransformers>=0.2.0

# API and server dependencies
fastapi>=0.104.0
uvicorn>=0.23.0
websockets>=11.0.0
pydantic>=2.4.0
python-dotenv>=1.0.0

# Database and memory system
sqlalchemy>=2.0.0
alembic>=1.12.0
faiss-cpu>=1.7.4  # Use faiss-gpu for GPU support
sentence-transformers>=2.2.0

# Utility libraries
tqdm>=4.66.0
requests>=2.31.0
aiohttp>=3.8.5
python-multipart>=0.0.6
loguru>=0.7.0
aiofiles>=23.2.1

# Optional optimizations
xformers>=0.0.23
bitsandbytes>=0.41.1
